refactor: Source-distributed `MapboxMapController` & `MapboxLayerHost`
The goal of this is to fix an issue where the pre-compiled `MapsGLMapbox.xcframework` wasn't linking properly across multiple Mapbox versions in client applications, especially relative to the changes Mapbox made in 11.6.x vs. 11.5.x.

The solution is to move most of the functionality out of `MapboxMapController` & `MapboxLayerHost` into (new) parent classes, and distribute the source for them instead, such that the client application compiles that source code against whatever versions of MapsGL and Mapbox they're using.

* MapsGLMapbox target's Sources now have `MapboxMapController` & `MapboxLayerHost` subclasses, along with a handful of small utility extensions for `MapboxMaps.CustomLayerRenderParameters`, `MapBounds`, `MapboxMaps.StyleManager`, & `MapBounds`.
* Removed the MapsGLMapboxWrapper target, and changed MapsGLMapbox to be a normal source `.target(…)` (instead of a `.binaryTarget(…)`).  Also removed the related `xcframeworkChecksums.mapbox`.